+++ packages/core/src/config/config.ts
@@ -240,50 +240,4 @@ export const createConfig = <
   },
 >(
   config: TConfig,
-): TConfig => {
-  // convert to an easier type to use
-  const contracts = config.contracts as Record<
-    string,
-    Contract<string, AbiEvent[], string, AbiEvent>
-  >;
-
-  Object.values(contracts).forEach((contract) => {
-    if (typeof contract.network === "string") {
-      // shortcut
-      const network = config.networks[contract.network];
-      if (!network)
-        throw Error('Contract network does not match a network in "networks"');
-
-      // Validate the address / factory data
-      const resolvedFactory = "factory" in contract && contract.factory;
-      const resolvedAddress = "address" in contract && contract.address;
-      if (resolvedFactory && resolvedAddress)
-        throw Error("Factory and address cannot both be defined");
-    } else {
-      Object.entries(contract.network).forEach(
-        ([networkName, contractOverride]) => {
-          if (!contractOverride) return;
-
-          // Make sure network matches an element in config.networks
-          const network = config.networks[networkName];
-          if (!network)
-            throw Error(
-              'Contract network does not match a network in "networks"',
-            );
-
-          // Validate the address / factory data
-          const resolvedFactory =
-            ("factory" in contractOverride && contractOverride.factory) ||
-            ("factory" in contract && contract.factory);
-          const resolvedAddress =
-            ("address" in contractOverride && contractOverride.address) ||
-            ("address" in contract && contract.address);
-          if (resolvedFactory && resolvedAddress)
-            throw Error("Factory and address cannot both be defined");
-        },
-      );
-    }
-  });
-
-  return config;
-};
+): TConfig => config;
